Move the block widgets to their own module
